refactor(tests): split test_persona_runtime.py into focused modules
The 3 398-line monolith violated the 500-line file-size policy that applies
to all source files. This commit:

1. Removes the blanket `tests/**` exclusion from `file_size.py` so Python
   test files are now subject to the same 500-line limit as production code.
   Pre-existing oversized test files are grandfathered; new test files must
   stay under the limit.

2. Deletes `test_persona_runtime.py` and replaces it with 10 focused modules
   (all ≤ 500 lines), plus shared helpers extracted to `_persona_test_helpers.py`
   and a `conftest.py` for the autouse registry fixture:

   - test_persona_state.py          — Mood, PersonaState, render_behavior, imports
   - test_llm_persona_agent.py      — _LLMPersonaAgent core behavior
   - test_persona_agent_factory.py  — factory, energy, event formatting, limits
   - test_persona_agent_validation.py — state validation, tools, sub-agent
   - test_persona_agent_errors.py   — memory close, corrupted state, MAX_TOKENS
   - test_persona_timeouts.py       — event/tick timeouts, minimal config
   - test_memory_injection.py       — _inject_memory_context core behavior
   - test_memory_injection_advanced.py — truncation, stale-section clearing
   - test_memory_instructions.py    — system prompt instructions, delimiters
   - test_memory_notes.py           — note injection, namespace, user identity

All 197 tests pass; `file_size.py --strict` exits 0.
