+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-25  
-> **386** tracked files
+> **397** tracked files
 
 ## Summary
 
@@ -10,7 +10,7 @@
 | `internal/` | 93 |
 | `docs/` | 81 |
 | `agents/` | 77 |
-| `tests/` | 45 |
+| `tests/` | 56 |
 | `(root)` | 20 |
 | `scripts/` | 17 |
 | `.github/` | 15 |
@@ -26,7 +26,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 133 |
+| `.py` | 144 |
 | `.md` | 97 |
 | `.go` | 90 |
 | `.yaml` | 25 |
@@ -477,6 +477,8 @@
 │       ├── __init__.py
 │       └── python/
 │           ├── __init__.py
+│           ├── _persona_test_helpers.py
+│           ├── conftest.py
 │           ├── test_agents.py
 │           ├── test_base_agent.py
 │           ├── test_base_handle.py
@@ -488,11 +490,20 @@
 │           ├── test_episodic_memory.py
 │           ├── test_event_dispatch_tick.py
 │           ├── test_llm_client.py
+│           ├── test_llm_persona_agent.py
+│           ├── test_memory_injection.py
+│           ├── test_memory_injection_advanced.py
+│           ├── test_memory_instructions.py
+│           ├── test_memory_notes.py
 │           ├── test_memory_tools.py
 │           ├── test_participant.py
 │           ├── test_permissions.py
+│           ├── test_persona_agent_errors.py
+│           ├── test_persona_agent_factory.py
+│           ├── test_persona_agent_validation.py
 │           ├── test_persona_nickname_generator.py
-│           ├── test_persona_runtime.py
+│           ├── test_persona_state.py
+│           ├── test_persona_timeouts.py
 │           ├── test_registration.py
 │           ├── test_relationship_memory.py
 │           ├── test_relationship_memory_user.py
