+++ tests/unit/python/conftest.py
@@ -1,0 +1,10 @@
+import pytest
+
+from agents.tools.registry import clear_registry
+
+
+@pytest.fixture(autouse=True)
+def _clean_registry():
+    clear_registry()
+    yield
+    clear_registry()
